Fix envsetup.sh error when sourced
A command used in this script could throw an error saying that invalid
options were provided. This is likely caused by the variable that is
used as the parameter of the command, of which its content may start
with a dash. By adding '--' before any options, it tells the command
that anything after the dashes should be interpreted as parameters, not
as options.